+++ index.ts
@@ -1,3 +1,9 @@
+import {
+  FormattedDataUnit,
+  Purchase,
+  RawPurchaseDataType,
+} from "./types/types";
+
 import rawData from "./purchase-file.json";
 
 /**
@@ -7,48 +13,33 @@ import rawData from "./purchase-file.json";
  * key values is customer id
  */
 
-type RawPurchaseDataType = {
-  customerId: string;
-  date: string;
-  items: string;
-};
-
-type FormattedDataUnit = {
-  customerId: Array<Purchase>;
-};
-
-type Purchase = {
-  date: string;
-  products: {
-    food: string[];
-    drinks: string[];
-  };
-};
-
 const parseData = () => {
-  const structuredData: any = {};
+  const structuredData: Record<string, FormattedDataUnit> = {};
 
   rawData.forEach((element: RawPurchaseDataType) => {
     const allItems = element.items.split(",");
     const food: Array<string> = [];
-    const drink: Array<string> = [];
+    const drinks: Array<string> = [];
 
     allItems.forEach((item: string) => {
       if (item.includes("FOOD")) {
         food.push(item);
       } else {
-        drink.push(item);
+        drinks.push(item);
       }
     });
-    const purchase = {
+    const purchase: Purchase = {
       date: element.date,
       products: {
         food,
-        drink,
+        drinks,
       },
     };
-
-    structuredData[element.customerId] = [purchase];
+    structuredData[element.customerId] = {
+      purchases: [purchase],
+    };
   });
   return structuredData;
 };
+
+console.log(JSON.stringify(parseData()));
